Fix screen scrolling and some other changes
Should have committed the screen scrolling fix separately, never mind.
I'm still not sure this is quite right - the initial screen seems to
pause unnecessarily early for a "More" prompt (albeit invisible) which
doesn't seem right, especially as making full use of the screen is a key
feature of ozmoo. I started poking at the other stuff to try to figure
this out but got sidetracked.