Fix endianness of Acorn kernel_readtime emulation
I'm sure I was reading the C64 programmer's reference guide right
("accumulator contains the most significant byte") but there's no
arguing with results and the ozmoo code clearly a) works on a C64 b)
expects this order. This slows down freefall.z5 on Acorn to what looks
like pretty much the same speed (comparing side-by-side in emulators) as
the C64.

Freefall does actually work, it's just the keyboard input appears to be
case-sensitive and you need to enter lower case. Toggling CAPS LOCK on
VICE seems to give the same kind of breakage I see on the Acorn by
default. Need to think what, if anything, to do about this.