Tweak Acorn keyboard handling
Only lightly tested but seems to work. I think the case-sensitivity of
freefall.z5 is just part of how it is. Obviously we could add a
conditional "force lower case input" option to the interpreter (since
it's built specially for each game) or the loader could force CAPS LOCK
off as a compromise (though nothing would stop the user turning it back
on), but since it's just a novel demo and upstream hasn't felt the
need to add interpreter features for it I won't worry about it yet.