Improve Acorn text printing logic
We no longer spam the OS with VDU 31 codes all the time.

I've also fixed a bug - I was trying to use X to record whether we
needed to clear a pre-scroll text window or not, but for the status bar
X was being corrupted and we were clearing a non-existent window
repeatedly. This was spotted by *SPOOLing the output and then *DUMPing
it.

I haven't tried to time it, I think the output is a little faster. Mode
7 flies, though it wasn't too sluggish before. The software scroll in
mode 6 is obviously hurting a bit, though this is nothing new.

Obviously I am still using a second processor (a Master Turbo as well),
but these enhancements - apart from being much more satisfying - will
probably be more relevant to performance on a standard machine, where
the same CPU has to handle the VDU processing as well as run the
interpreter.