More unsuccesful VMEM hackery
TBH I'm probably going to have to bite the bullet and use make.rb to
generate a config and load that at runtime. At the moment I'm trying to
coax an (inefficient, but that would be fine, as good for testing)
initial version to run with nothing preloaded for simplicity.