Disable vmem_cache on Acorn
I've had a read through the code and I think I understand it; we don't
need it. I think this frees up about $600 bytes (not checked exactly),
$400 bytes for the cache and probably the rest for code we've been able
to delete and "rounding" for alignment.

There's a 5.8% reduction in the time taken for the benchmark, which is
nice, presumably due to a mixture of not using the cache (I had
"incorrectly" left first_banked_memory_page at $d0 on VMEM builds) and
having extra RAM so we can avoid a bit of disc access.

This has been pretty epic, what with finding the bugs in the previous
commits and also failing to ensure story_stage was 512-byte aligned at
first.

(I poked at the source for HH on github in order to figure out that the
song you need to play is random. It was fairly obvious where things went
wrong from watching the benchmark auto-play enter commands and seeing
them fail, so I started poking around at the piano-related code.)