Fix disk read bug, get rid of PRELOAD file, misc related tweaks
Benchmark passes.

We no longer have PRELOAD and therefore no longer need the overlapping
files trick, which is particularly undesirable now a user might be
saving onto this disc. This is negligibly slower as per sf-notes.txt.

Turns out I wasn't correctly checking for errors after calling OSWORD
$7F, this should now be fixed. This showed up by trying to use 4-sector
reads in the preload loop; I think this fails because we will try to
read across a track boundary (10 is not a multiple of 4), and without
checking for the OSWORD $7F error this meant the benchmark rapidly
failed as we had failed to read all the data.