Tidy up OSWORD $7F code
This seems a lot neater, and it's shorter too. Tested with a few
different DFS versions and seems to work fine (benchmark completes OK).

The 16-byte tube copy mentioned in the now-deleted comment was a red
herring - it's only a problem if the data being read is being placed
just after the control block, and here it isn't.