Alter how we default to S directory
By making the loader do it, if the user changes this via * commands
during a game that change isn't lost on a restart. This isn't a big
deal, but it seems a bit nicer.

(The binary will be slightly smaller, but the code was discarded after
startup so this isn't a significant improvement in memory use.)