Check for spaces in filenames
I'm not sure about this. If nothing else, maybe we shouldn't re-prompt
for a filename (if you enter an invalid filename, an error occurs on
save/restore and the save/restore is aborted). However, let's stick with
it for now. This may not actually be my problem, but I've had a few
saves which I "thought" I'd called something else save as "SAVE", and
although I don't think this is due to a bug in the code it is worrying,
so let's see if this helps. Once I have convinced myself these "SAVE"
files are just user error I could always get rid of this check if I
wanted.