Remove TODO
If we used scratch space for this control block we'd need 18 bytes of
code to zero it and populate the pre-populated word, and the control
block is 18 bytes. So we're no better off for memory if we use scratch
space, and with scratch space the code is a bit more complex and we need
to worry about trampling over scratch space some other code is using, so
the current approach wins.