Don't use zero page outside $00-7F
This will be necessary for a non-second processor version, and it
doesn't really hurt on a second processor - the variables moved outside
zero page due to lack of space aren't performance critical, and as they
only moved into $400-800 they aren't stealing a few bytes of space which
could be used for code or game data.

The benchmark completes OK, so with luck I haven't broken anything.