Fix bug in printing at bottom right of screen
When this happened we would fail to reset the text window and output
"garbage" via the pha:pla:jsr oswrch (although "luckily" the garbage
would either be a 0 in non-reverse video mode or vdu_reset_text_window
in reverse video mode). We would mostly get away with not having reset
the text window, because the OS "clamps" if you try to position the
cursor beyond the bottom edge of the window. I think it would be
possible for things to go wrong if e.g. the status bar were redrawn
while this unwanted text window was still in effect, but most of the
time we'd get away with it because something else (e.g. a newline where
a line of text didn't reach quite to the right hand edge and we scrolled
explicitly) would use a text window and reset it.

I found this by looking at the code while experimenting with hardware
scroll optimisations, I haven't actually *seen* it go wrong.