Add hardware scrolling support
This is much faster - see the benchmark results in text.asm - but the
status line flickers a lot as a result. I'll leave it on for now so I
can test it, but I don't know if it's a good idea in the long run. Then
again, if I start supporting mode 0, software scrolling is going to be
even slower.