More SWR experiments
I had actually modified the wrong version of one routine in vmem.asm but
was getting away with it as the RAM bank was left paged in all the time.
I've now hacked things so we page it out outside a few key routines -
this would have showed the problem, but I found it by reading code while
adding the hack.

The benchmark does complete like this, so this suggests I have a lot of
flexibility about paging, I don't need to leave banks paged in if
I'm willing to pay the modest price for switching frequently.