Fix relocation bugs
The benchmark now completes on B+, Master 128 and Master Turbo. Using
the b-em debugger it looks as though the code is running at the right
address on each machine.

I haven't implemented the B-with-no-shadow-RAM support yet, so that
obviously won't work.

Testing in BeebEm on Integra B it fails with a disc read error during
load, whether I use the default Watford DFS 1.44 or Acorn DFS 1.20. That
"should" work already.