Add --fake-read-errors to make and fix revealed bugs
As per an SFTODONOW comment, the OS helpfully pages the current language
back in instead of the bank we wanted when an error occurs, so we need
to fix that up.

There was also a lurking hardware scroll bug in ACORN_NO_SHADOW mode
when disc errors during initial loading caused the protective text
window to be removed before we executed the vdu_cls to clear the screen
just before starting the game.

As far as I can tell everything works fine with these changes.