Thoroughly hacky SWR optimisation
The benchmark passes so want to commit this before fiddling further.

The idea here is to avoid constantly paging in and out during
instruction decoding - following some hacky profiling of the benchmark -
and I this may well work, but I've only just managed to get it to to not
crash so the code is a hack and needs to be tidied up before I can
determine the performance benefits, if any.