Remove TODO
I've had a look at the profile output, this is not a common case. It's
mildly annoying but arranging for read_byte_at_z_address to know it
shouldn't bother complying with the invariant "the bank
z_pc_mempointer_ram_bank should be paged in" in this case is not
trivial, particularly given the other conditional assembly going on. (If
we set a flag and check that, we're then paying for that flag check all
the time, which probably isn't a net win. So we'd need to separate the
code paths somehow.)