Factor out vdu_goto_xy calls
It doesn't save a lot of space, but I think the resulting code is
cleaner. There is technically some risk of breakage here as some
routines now corrupt X and Y when they previously didn't, but I've
reviewed the code and given it a quick test and I'm fairly sure it's
fine.