Tweak TODOs
The way I do the "experimental" builds from low addresses upwards mostly
negates the concern about them causing a die() when we would have tried
something else if only we'd been clever enough. I am not saying this
won't ever change, but it's not as big a deal as I thought when I bashed
that TODO in.