First cut at ADFS "nice" save/restore/error handling
After all my agonising in sf-notes.txt, this actually isn't that big a
deal and it seems to work pretty well. If you put a DFS floppy in the
drive you're likely to get stuck with an unrecoverable "Bad FS map"
error and no way out, but otherwise this seems to work pretty well based
on fiddling around on b-em. The code is a bit grotty but the basic idea
is sound, I think.