Make loader check PAGE
This requires minor restructuring of make-acorn.py because we need to
build the executables first so we know their load addresses before we
generate the loader and tokenise it. It's not a big deal but I've
retained the "natural" ordering of the files on the disc (boot then
loader then findswr then ozmoo binaries then data) to try to reduce head
movement.

This might also be helpful in supporting builds where one or more of the
current three variants isn't available.