Improve loader mode menu for Electron
I'm not sure this is perfect - I hate this code, I really do, I just
don't have the comfort with BASIC <=4 - but the actual user behaviour of
the menu is probably correct and it now does something sensible in mode
6 on the Electron, although this would only be relevant if you had a
second processor anyway.

The loader as a whole is still not properly Electron friendly but this
is a step in the right direction.