Re-implement adjust_dynamic_memory_inline
Very lightly tested (a bit of stepping through code in the debugger on
a 144K SWR model B with Z5 THHGTTG) but does seem to work.

In many ways this is cleaner on the new model; I do plan (see SFTODO) to
pre-age any vmap entries which have become redundant thanks to the
growth in nonstored_blocks so they're re-used ASAP, but we no longer have
to fiddle around with the vmap entries to make this work because the
load_suggested_pages code will just load the right things in the right
place - yes, it will (even once I deal with the SFTODO) load unwanted
duplicates of the "extra" nonstored_blocks (assuming they're in the
vmap, of course, though this is likely) but that's not a huge loss. (I
really don't want to be working out which game blocks are not going to
be loaded in the initial load so I can substitute them into the vmap as
part of the nonstored_blocks-growth optimisation; the code would hardly
get any testing, so I want to keep it as simple as I can, and the only
loss is a bit of extra loading on startup and then an extra read the
first time those blocks get populated with something useful from disc.)