Generate shuffled block list in vmap
This is obviously silly as a whole but a) the idea of having an ordered
list of blocks will obviously be useful once I have preopt support b)
this sets the scene for adding a sort to load_suggested_blocks. At the
moment the drive thrashes about reassuringly during the initial load but
the benchmark completes (tried B+128K SWR and Master Turbo) as I'd have
expected. You can also see the shuffle using --trace-floppy.