Fix bug in recent changes shown by Electron
Without this the Electron hung on startup. The preceding loop will
leave the last bank paged in; in practice on the BBC we get away with
this (though I am not sure it's absolutely guaranteed) but on the
Electron we don't because all dynamic memory is in sideways RAM and
we have broken the invariant that the dynamic memory bank should be
paged in by default.

(I wanted to commit just this fix, but I also accidentally committed
some minor comment tidying; never mind.)