Enable dynmem growth on Electron
I haven't tested this - I don't know if an Electron *can* have enough
sideways RAM to trigger this - but I can see no reason at all for it not
to work. (I wanted to get it working before thinking about this, but the
doubt was mainly an artefact of an overly-complex scheme I had in mind
before I realised I could do it as the last couple of commits have done
- I didn't realise my implementation easily allowed a gap in virtual
memory backing RAM around the screen, so I was going to start the
contiguous virtual memory cache at what is now extra_vmem_start and put
some "do not use" blocks in to cover the screen RAM and the dynamic
memory. With that scheme there was a lot more fiddling around with vmap
and related variables and this optimisation may or may not have worked,
plus on an Electron quite a few vmem blocks would have been unusable.)