Rename ACORN_ELECTRON to ACORN_ELECTRON_SWR
The tube build works on the Electron but currently has no changes for
the Electron. I think it's clearer to indicate that the ACORN_ELECTRON
macro tells us "we are building Electron SWR", especially in case I need
to pass a "build for Electron but not necessarily for SWR" flag to some
binaries at some point.

Do wonder if it should be ACORN_SWR_ELECTRON not ACORN_ELECTRON_SWR
given other SWR stuff is ACORN_SWR_* but it's really about the Electron
not just about the SWR.