Replace ACORN_SAVE_RESTORE_OSFIND with ACORN_SAVE_RESTORE_OSFILE
The sense is opposite; as per now-deleted SFTODO, it feels better for
the slow-but-safe option to be the default if nothing is defined. In
reality this isn't a big deal (the code defines this internally, it
isn't set by the build script - well, it was for the Electron, but I
am not sure it needed to be and it isn't now anyway), but still.

Generated .ssd is identical for the benchmark before and after this
commit, so this is probably correct.