Tweak Electron loading screen
With the coloured background the asymmetry (which is invisible in mode 7
or mode 6 with a black background) caused by the dummy control codes in
column 0 is more obvious, so even things up by not extending the
horizontal rules into column 39. I think this looks a bit better. It
would probably be best to just use the full width of the screen, but
that would be a bit more fiddly while maintaining mode 7 support.