Pull Acorn disc code into acorn-disk.asm
disk.asm is now almost identical to upstream; it only differs in that it
defines WANT_RESTART to tell the VM instruction table whether to include
z_ins_restart or not.

Acorn HH and THHGTTG disc images are identical with and without this
commit, so it should be correct. (This wasn't true at first, as I had
overlooked WANT_RESTART being important outside this file, which shows
the value in doing this kind of binary comparison.)