Fix bug in restore using OSFILE
The load address in the OSFILE block wasn't being updated before each
OSFILE load call. This meant the first restore was always fine but it
could "corrupt" the load address if the file being restored hadn't been
saved with exactly the same build of Ozmoo. (I think I experienced this
by restoring second processor saves onto a sideways RAM build.) We got
away with this if the file being loaded had been saved by the same build
of Ozmoo, so casual testing would miss this.

The fix about not using an outdated mempointer_ram_bank was something I
thought was causing this crash at first; it wasn't, but I think there
was a risk here so I've left the fix in even though I haven't seen this
problem occurring.