Fix max_age in make-acorn.py, tweak naming/comments
The tube caching stuff is forcing me to think about this more and I now
realise I was doing this wrong. make.rb won't generate a timestamp newer
than $e0, but this code would, so preloaded blocks could actually be
considered newer than a block just touched at runtime. In practice this
was probably not a huge problem, but it wasn't really right.

I've changed "age" to "timestamp" because "age" makes me think "high
value=old", whereas with timestamp I think "high value=recent", and the
latter is how it actually works (because vmem_tick is incremented as
time passes, as you'd expect).