Rework tube host cache load
We now use the new timestamp hint to pass the ordering information
present in the initial vmap through to the host cache.

(The timestamp hint is not wanted when we're offering blocks to the
cache during normal play; there the host cache will be marking newly
offered blocks with the most recent possible timestamp. It's only during
the initial load that the hint is wanted.)