Fix mostly irrelevant bug, add TODOs
In practice this problem wouldn't occur, because it only mattered if we
didn't need all the local RAM to hold the game, and in that case the
build system would switch to a non-VMEM build, but it made me uneasy and
this is discardable init code so let's fix it. That way if I've got my
analysis wrong and the previous sentence isn't true we're less likely to
get a nasty surprise in future.