Fix brokenness in previous commit
FFS... I think the issue was this was working by chance and offsetting
things by an odd number of bytes with the "!fill 3" was triggering the
broken case.

I *think* this might have worked fine on Master for some reason, but BBC
B systems tended to break (very late in the benchmark).