Fix vmap_sort_entries corruption bug
At least I noticed; this meant the initial vmap was not what the user
specified when using preopt because the sort was sorting too little or
too much. I suspect this made no difference when not using preopt
because the vmap would already be sorted.

There's no need to try to squash all the initialisation zero page use
into zp_temp, because this is initialisation code we can safely use
things like vmem_temp as I have here.