Tweak splash screen preloader code
This is mainly tidying up, but it also fixes a couple of low-probability
glitches:

- We no longer run the SPLASH executable if we're on a second processor
but the host OSHWM is too high.

- The "third party shadow RAM &3000 onwards corruption possibility"
handler (which is only relevant if the splash screen is in a mode other
than 0-2 anyway, because in modes 0-2 the previous check of 'will the
loader fit in this mode?' does everything necessary) now checks
explicitly for the Integra-B, because it fakes the host OS version.