Experimental optimisation
This works (at least as far as running benchmark in smalldyn mode goes)
but the saving isn't as big as I'd hoped. Profiling shows about half of
calls to read_next_byte go through the "fast path" and some formerly-hot
instructions in read_byte_at_z_address no longer are, so this is working
to some extent.

The improvement is tiny though, and while it just might be possible to
optimise things a little more (the two instructions checking the flag at
the start of read_next_byte are hot, and self-modifying code might
shave a few cycles off them, plus the clearing of the flag at the start
of read_byte_at_z_address is slow as it uses "lsr zp" and is a hot
instruction), my *suspicion* is that even with those things tweaked the
saving is not going to be enough to justify the (modest) extra
complexity. I'll sleep on it anyway.