Basic hacky memory hole working
This code is super-hacky but the benchmark completes and looking at the
page (remember it's just a one page hole at this point) at $7c00 in the
debugger afterwards it still contains the MOS startup banner.

I really need to have a proper think about all sorts of corner (and not
really even that corner, just "not this one case") cases, but I feel
happier for having this basic proof of concept working as a starting
point.