First cut at supporting private 12K on B+
Seems to work fairly well.

The code change looks bigger than it is because some of the RAM
calculation code was executing *after* dynmem_load_loop, which apart
from being pretty illogical meant that scratch_page was getting trampled
on and so some RAM calculation data I was trying to held there was being
lost.