Fix tube build after previous fix
Before the previous fix this wouldn't have mattered, it was (I think)
correct if unintentionally so, because a tube build would never have
ACORN_PRIVATE_RAM_SUPPORTED defined.

As can be seen in sf-notes.txt I've done some timings and things do seem
to be performing roughly as expecteds, so with luck this fix is correct.
I can only assume the reason I didn't notice earlier was that (through
luck?) the bug on B-no-shadow tended to cause sideways RAM to be
under-estimated, but I've not tried to dig into this too much.