Zero the history buffer on startup
This isn't necessary on Commodore as it's naturally zero-initialised as
it's part of the main executable, but on Acorn it may now live in page 4
instead.

The problem manifested itself by typing a few commands (not typically
long enough to fill the history buffer), then doing up-down-down, at
which point gibberish would be entered instead of the second down doing
nothing. Presumably because the exactly behaviour depended on the
contents of page 4 this couldn't be reproduced every single time, but it
seemed to happen quite easily once I noticed the pattern.