Tweak vmem cache handling
The code is still a mess, but this moves it towards being a little
cleaner. Benchmark completes OK in mode 7 on M128 and B+64K. The new
code is shorter and has a "more obviously correct" (if we assume it is
valid to corrupt X, which I think we are already heavily relying on)
"branch always" assumption. (Famous last words...)