Optimise away vmap_used_entries most of the time
Lightly tested and seems to work. This should shrink the code - both
through removing a few code fragments, and through using a zero page
address in places where we previously used a non-zp address - and
similarly may offer a (probably small and not yet measured) performance
improvement.

I was tempted to make vmap_used_entries and vmap_max_entries refer to
the same memory address, but doing it this way is probably much less
confusing in the long run. (Making them aliases would also not
automatically find places where we store to both, which would be
redundant at best and broken at worst.0