Put caching in make_ozmoo_executable()
This is simpler, I think; we don't need caching for all executables and
having it outside the Executable/OzmooExecutable classes feels cleaner.
This also fixes the odd build problem noted in the removed SFTODONOW
comment, which was almost certainly due to the cache key not being
complete in the older implementation.