Support basictool for BASIC tokenisation/packing
beebasm is still supported, but if both are available we prefer
basictool as ABE's pack is so much more powerful than our own ad-hoc
cruncher.

Note that a small tweak had to be made to the loader to avoid ABE's
line-concatenation breaking things. Falling through into
PROChighlight_internal would fail to execute any statement concatenated
on after the DEF PROChighlight_internal, which meant that pressing the
left arrow (for example) to move off the default mode 7 corrupted the
display. Explicitly calling PROChighlight_internal isn't a big deal and
avoids this quirk. The BASIC loader is about 25% smaller with ABE's
pack, so this is well worth it.