Tweak TODOs
The statement that with USE_HISTORY the INSV handler would allow *FX4,0
to be forced regardless was wrong - the INSV handler respects *FX4,1 and
doesn't allow split cursor editing in that state. I just tested this and
it means --no-cursor-editing does have meaning with history as well as
without it.

This isn't to say there isn't an argument for changing things, but I
think what we have is actually fairly sensible and I'm not tinkering
with it right now anyway.