Tweak loader private RAM support
Integra-B benchmark mode 7 with 32K SWR in banks 4 and 5 breaks during
loading, but it does the same before this commit, so might as well get
this committed before investigating further. Edit: That Integra-B config
works fine on BeemEm, so I'm going to put that down to b-em. I'll try to
sort this out, but it's not a problem with my code AFAICS.