Don't determine game size at runtime
We were already using the game size at build time in various ways
anyway. This simplifies the code quite a bit - it was discardable init
code so the space saving isn't a big deal, but it still added to the
complexity of following the code, particularly with things like the
double-sided hack meaning the game size wasn't always determined
precisely.

Lightly tested and seems to work OK.