De-macro-ise Acorn init code
This is a work in progress but the macros made it really confusing to
follow the code and were a nightmare to debug because acme report files
don't show the generated assembly language and addresses of instructions
for macros.

This seems to work as well as the previous commit - normal tube is still
broken, but other stuff I tested before still works.