Fix --no-dynmem-adjust, more code folding tweaks
I think there was a misplaced closing brace for the --no-dynmem-adjust
case before - the previous commit built a broken benchmark (tested on
M128) with --no-dynmem-adjust - and this one seems fine. It is possible
I've got confused.

This or something else had lead me to over-simplify the code folding
comments, but I think I've improved them now, and I think doing this is
helping to find problems like this --no-dynmem-adjust one.