Require basictool 0.06, refactor tool version checking
basictool 0.06 needs "-s" to strip spaces and this won't be recognised
by earlier versions; we *check* but we don't bother accommodating old
versions.

This will refuse to use a tool if it can't determine the version, which
is a change in behaviour; I don't think this is a big deal.