Always use 512-byte blocks with progress indicator
I think this is less confusing overall. Lightly tested, HH progress
indicator works fine on a Master 128 and arc3d progress indicator works
fine on tube (which tests non-VMEM case).

First use of dpage in anger. I might ditch the term and talk about "vmem
blocks" everywhere, but I'm a little hesitant when that's arguably
incorrect. Here for example we are loading 512-byte blocks from disc,
but the first N of them are *not* vmem blocks in the sense they are
handled by the vmem subsystem, although if we simply define "vmem block"
to mean "512 byte block" not "block handled by vmem subsystem" it's
fine. The question is what implicit definition will be less confusing.