Hacky but probably correct-ish SWR hole code
The benchmark completes on an Integra-B with 64K sideways RAM, and it
doesn't hang on BREAK afterwards. Uncommenting the !if 0 block of code
does trigger a failure in the experimental "assert the hole location has
been calculated correctly" code; I might well delete this now I've
reworked the calculation but I need to think about it and perhaps come
back to this fresh.