Tweak comments/labels
The comment might be stating the obvious, but I did suddenly think "I'm
not checking sideways_ram_hole_start != sideways_ram_hole_start_none",
so it's not *that* obvious. sideways_ram_hole_start is *how* we prevent
the trampling from occurring, but the trampling should *never* occur -
and the latter is what we're checking here.