Remove experimental sideways RAM hole verify code
I think this works, but it feels convoluted, I don't like having to do
this late enough that it's safe to call convert_index_x... (because
everything necessary has been set up) and the edge case ("bcs +") means
it might not always fire. With the debug code to help catch problems
directly in convert_index_x... - which is a lot more obviously correct -
I don't think this is necessary.