Semi-experimental tweak of sorting/initial load
This has been lightly tested and seems to work. I think - possibly after
a little more polishing/comment tweaking - this is cleaner than what we
had before. We're no longer sorting 0 dummy entries down into the early
part of the vmap and then removing them afterwards, and I think this
avoids a couple of probably-OK-but-I-was-dubious corner cases.