Add --nfs-install-only build option
Thanks to Ken Lowe for prompting this. It's a bit crude at the moment
but will allow building an ADFS-style game but putting the files on a
DFS disc image. They won't work from there (although it is possible to
manually hack the loader to make it mostly work, and I did that to test)
but can be copied onto NFS without requiring an ADFS-capable station to
do the install from.