Avoid sideways RAM voodoo pages, minor tweaks
I don't believe we need to page in the default bank of sideways RAM
during initialisation; perhaps we did at one point. So remove the
voodoo-ish code and replace it with something which will "prove" we
don't need to do this.

A few other minor tweaks also in this commit.