Don't alias vmap_index in host cache load
It's nice to have the symmetry of .from_index and .to_index but I don't
like the aliasing, it feels error prone. We weren't simply "re-using"
vmap_index's memory for .to_index, we relied on - given that's how
load_blocks_from_index is told which block to load - the two sharing an
address. I'd rather make that explicit.

I suppose it would be possible to have to_index at a random memory
location and deliberately copy it to vmap_index when loading a block,
but I note the second loop already used vmap_index "raw" so I don't
think this is too confusing.