Remove outdated comment
I just tested the benchmark in MAME with an Acorn 6502 copro and it
completed fine. I didn't measure the performance (I'll do that sort of
thing later in the tidying process anyway) but the cache executable
would be run and used and the fact it's not crashing because it doesn't
work on the Electron is the key point. As per the TODO, I'm 99% sure I
actually tested this ages ago anyway.