Tweaks re game being a multiple of 512 bytes
First off, note that hardly any of the complex code in
acorn-init-preload.asm executes in the non-VMEM case - the folding does
help to make this more apparent. So most of the apparent complications
in a non-VMEM build caused by data_start not being dpage aligned are
irrelevant.

We also were already implicitly assuming the game was a multiple of 512
bytes even on non-VMEM because the load loop would fail to terminate
correctly (I think we'd load twice as much as we should, but I may be
wrong) anyway.

In practice all this means we could just double page align data_start
even on a non-VMEM build and we'd be no worse off at all, but I won't
force this when it does no harm. It's possible something will change in
the future which makes this useful.