Review/tweak TODOs
I went over all the SFTODOs in acorn.asm, acorn-init-preload.asm and
acorn-init-stack.asm and turned the ones which looked important or
simple-and-tempting enough into SFTODONOWs as part of this.

I've been intermittently (on sight) pruning out the temporary SFTODO
labels I used to need when forced to debug the code in the huge
initialisation macros; now the code no longer lives in macros these
aren't necessary as the address of any instruction can just be read from
the acme report output.