Fix overrun when clearing bulk_clear_{start,end}
There was an implicit assumption that the low byte of bulk_clear_start
was 0, which wasn't true in practice. In practice this didn't really
matter as the following memory is also part of language workspace and
doesn't contain valuable data at this point, but the code was still
confusingly wrong.

I do wonder if I should use (zp),y addressing here instead of
self-modifying code, but there's not that much in it. (If I *did* use
(zp),y I would want to do this before clearing zero page, in order to
leave zero page all-zero afterwards.)