Tweak runtime info code/comment
This still seems to work. Calling deletable_screen_init_1 was harmless
(this is discardable init code and it's debug code only) but it and the
comment were a bit confusing, because we *already* called it earlier
anyway. The key thing is that we need to call *streams_init* (and this
isn't necessary for s_printchar to work, so there shouldn't be a problem
reporting errors during startup via BRKV).