Assert vmap_meaningful_entries >= 1
This has never not been the case, but asserting helps rule out the
possibility of a confusing incorrect loop when sorting the vmap. This
should be guaranteed to be true because we ensure we always have at
least one block of vmem whatever happens, but there's enough dynamic
fiddling around with the vmap that it's possible there is a bug here so
we might as well check.