Tweak upper memory threshold in assert
With the new expanded discardable runtime init code, a benchmark build
of OZMOOSH strays up past 0x7c00 (but not 0x8000). This is fine (at
worst it would corrupt the loader screen, which is ugly but harmless,
and we'd notice and fix it) and in practice it doesn't even do that
because we're in shadow mode. This seemed a better fix/workaround than
altering bbc_max_start_addr.