Use hard-coded *10 in init_read_text_timer
This is what upstream does and it allows the mult16 subroutine to be
removed. This saves 28 bytes (inferred from program_end before and after
this change; I haven't sat down and counted based on the code).

I ran builds of freefall.z5 before and after this change side by side
and they appear to be running at the same speed.