Tweak code and add TODOs re 40 column support
I am not sure --only-40-column was properly implemented, but I may be
wrong.

Note that SUPPORT_80COL is tested using !ifdef not by its numeric value,
so the previous version of the code was wrong before this tweak. This is
probably not the *only* lurking --only-40-column related problem.