Use immediate trick for is_buffered_window
This frees up a zp location which I haven't yet reallocated, and appears
not to harm performance in and of itself. (We hardly ever write to this,
so extra cost there is irrelevant, and one of the two uses is a cycle
faster as it's an immediate operation, and the other use gets a cycle
slower but I think it only occurs after the first use, so we come out
even at worst. But this is not a rigorous analysis - the timing test is
what makes me reasonably happy with this.)