Move vmap_c64_offset into zero page
This is a tube-only variable and we have zero page spare (currently) on
tube. I am disappointed the benchmark didn't show a performance
improvement but since it's not forcing anything out of zero page I don't
see how this can be a bad idea. (I haven't tried to confirm it, but I
suspect the improvement is there but masked by accidental alignment
changes.)