More second processor tweaks
I feel a bit bad doing this as the performance timings show a small
regression, but since we really are just moving stuff that would
previously be in low memory into zero page, I don't really see how it
can make things worse. Just possibly the history buffer is larger and
that slows the benchmark down, but even if so (and it's just
speculation) that would not really be a user-perceptible slow down. Also
vaguely possible this is due to random bad luck page boundary crossing
as the code gets shifted in memory.