Remove dead code
As per note in shadow driver source, for now I'm not generating an error
in this case, just reverting to screen only. The big advantage of this
is it means we should correctly support otherwise unknown shadow RAM at
least for screen only, instead of treating such RAM as a failure caused
by old Watford DFSes.