Fix bug in host cache shadow paging support
We are potentially accessing two different independent blocks of cache
during each OSWORD call - one to write an offered block to cache, one to
read a requested block from cache. We therefore need to undo paging in
the shadow RAM bank after *each* of those separate operations, not just
at the end of processing the OSWORD call.

I was misled by the way I handle sideways RAM paging, but there we
always page in the desired bank for each of the two operations (and it
doesn't affect operations which don't use sideways RAM), whereas we do
the shadow RAM paging only when necessary rather than explicitly paging
in main RAM when we want it.

Lightly tested (on Master) but seems to work.