Check host cache changes haven't made it bigger
We had 10 cache entries free on a BBC B with no shadow/sideways RAM in
mode 0 with alpha 45, and we still do with the current code. So all the
new shadow RAM logic isn't penalising systems which don't have any by
bloating the code and reducing main RAM free for caching.