Remove Commodore-specific code in screenkernal.asm
HH and HHGGSG seem to build identical disc images before and after this
commit, so I'm reasonably optimistic I haven't broken anything yet.

I may regret doing this, but it's hard to see the wood for the trees in
the (rather differently structured) Acorn code with all the Commodore
stuff in here. I can obviously still refer to the Commodore original
code separately, and it's not as if I expect scrollback to be
implemented on Acorn in the same way as on Commodore or anything like
that.