Copy and paste chunks of OS 1.2 into custom OSWRCH
I want to avoid calling directly into the OS so I can come up with a
driver that works on all BBC machine. For now I've just copy and pasted
big chunks of TobyLobster's OS 1.2 disassembly in, but I can now start
to tweak this code and optimise it for my specific use case.

Lightly tested and seems to work